backend/requirements.txt
fastapi
uvicorn
python-multipart
supabase
python-dotenv
pydantic
